backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlmodel==0.0.22
asyncpg==0.30.*
sqlalchemy[asyncio]==2.0.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
bcrypt>=4.0,<5.0
slowapi==0.1.*
aioboto3>=13.0,<14.0
python-multipart==0.0.*
pydantic[email]==2.*
pydantic-settings>=2.0,<3.0
httpx==0.28.*
